% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VoI.R
\name{voi_glm}
\alias{voi_glm}
\title{voi_glmn function}
\usage{
voi_glm(
  n_sim,
  thresholds = (0:99)/100,
  glm_object = NULL,
  mc_type = "bootstrap",
  data = NULL,
  pi = NULL,
  truth_formula = NULL,
  family = binomial(link = "logit")
)
}
\arguments{
\item{n_sim}{Number of simulations required.}

\item{thresholds}{thresholds at which EVPI is calculated}

\item{glm_object}{The GLM object representing the proposed model}

\item{mc_type}{any of "bootstrap", "Bayesian_bootstrap", or "likelihood"}

\item{data}{for EVPI calculations.}

\item{pi}{A vector of predicted values. Required only if data is provided. If a string or single number it indicates the corresponding columns in data}

\item{truth_formula}{Formula of the correct model. Its parameters will be estimated repeatedly using the data}

\item{family}{GLM family and link function}
}
\description{
voi_glmn function
}
